+++ src/Event/Fcl/fcl_misc.c
@@ -1143,7 +1143,7 @@ u32 FUN_003c94e0(void)
 // FUN_003C9570 NONMATCHING
 
 
-void FUN_003c9570(int param_1)
+void FUN_003c9570(u64 param_1)
 
 
 
@@ -2354,7 +2354,7 @@ void FUN_003cb100(int param_1)
 // FUN_003CB1F0 NONMATCHING
 
 
-void FUN_003cb1f0(u64 param_1)
+void FUN_003cb1f0(int param_1)
